requirements.txt
pip==20.1.1
flake8
tox
coverage
Sphinx
twine
pytest
pytest-runner
pyaml